requirements.txt
torch
torchvision
ftfy
regex
tqdm
pillow
numpy
scipy
jsonlines
requests
# open_clip_torch
transformers